+++ packages/xchain-terra/src/const.ts
@@ -1,0 +1,1 @@
+export const TERRA_DECIMAL = 6 // 10^6