+++ packages/xchain-terra/CHANGELOG.md
@@ -1,3 +1,22 @@
+# v.0.1.0-alpha.2 (2022-02-04)
+
+## Add
+
+- `TERRA_DECIMAL` #481
+- helper `isTerraNativeAsset`
+
+## Update
+
+- `xchain-util@0.6.0`
+
+## Fix
+
+- Fix chain in constructor #483
+
+## Breaking change
+
+- `getTerraMicroDenom` returns null for invalid denominations
+
 # v.0.1.0-alpha.1 (2022-02-04)
 
 Initial release